Fix Elasticsearch NotFound Error
Elasticsearch::Transport::Transport::Errors::NotFound
The above error is thrown when a deletion is made in the database but
the row had not been indexed into the cluster.

Since we don't care if the document is indexed at deletion time, there
shouldn't be an error thrown.  Catch the error and raise a warning
when trying to delete from the cluster.

Add specs for callbackable module
Increase gem version number